requirements.txt
# Core Google Cloud and Vertex AI dependencies
google-cloud-aiplatform>=1.38.0
google-adk>=0.1.0
vertexai>=1.38.0

# Authentication and configuration
google-auth>=2.15.0
google-cloud-core>=2.3.0
python-dotenv>=1.0.0

# Flask web framework and extensions
Flask>=2.3.0
Flask-CORS>=4.0.0

# Additional utilities
requests>=2.28.0
urllib3>=1.26.0

# Production server (optional, for deployment)
gunicorn>=21.2.0
Werkzeug>=2.3.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-flask>=1.2.0

# Code quality (optional)
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0

# Logging and monitoring
structlog>=22.0.0

# Security
cryptography>=41.0.0